+++ plugins/SunFlower.Le/Headers/Le/VxdHeader.cs
@@ -5,8 +5,8 @@ namespace SunFlower.Le.Headers.Le;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct VxdHeader
 {
-    public uint e32_winresoff;
-    public uint e32_winreslen;
+    public long e32_winresoff;
+    public long e32_winreslen;
     public ushort e32_devid;
     public byte e32_minor_ddk;
     public byte e32_major_ddk;
